feat(import): allow adjustment of speed of engine data fetching
related to #OPT-630

+++ backend/src/main/java/org/camunda/optimize/dto/engine/CountDto.java
@@ -2,13 +2,13 @@ package org.camunda.optimize.dto.engine;
 
 public class CountDto {
 
-  protected int count;
+  protected long count;
 
-  public int getCount() {
+  public long getCount() {
     return count;
   }
 
-  public void setCount(int count) {
+  public void setCount(long count) {
     this.count = count;
   }
 }
